Moved IHttpSender out from RestClient.
