+++ src/Libraries.Web/RestClientHelper/IRestClient.cs
@@ -10,17 +10,8 @@ namespace Nexus.Link.Libraries.Web.RestClientHelper
     /// <summary>
     /// Support for calling a REST service.
     /// </summary>
-    public interface IRestClient
+    public interface IRestClient : IHttpSender
     {
-        /// <summary>
-        /// The URI for the service. Separate methods in the service are called by using URL:s relative to this path.
-        /// </summary>
-        Uri BaseUri { get; }
-
-        /// <summary>
-        /// Credentials that are used when sending requests to the service.
-        /// </summary>
-        ServiceClientCredentials Credentials { get; }
 
         #region POST
 
@@ -214,49 +205,6 @@ namespace Nexus.Link.Libraries.Web.RestClientHelper
         Task DeleteAsync(string relativeUrl,
             Dictionary<string, List<string>> customHeaders = null,
             CancellationToken cancellationToken = default(CancellationToken));
-        #endregion
-
-        #region Send 
-
-        /// <summary>
-        /// Send a request with method <paramref name="method"/> to <paramref name="relativeUrl"/>.
-        /// </summary>
-        /// <param name="method">POST, GET, etc.</param>
-        /// <param name="relativeUrl">The Url relative to <see cref="BaseUri"/>, including parameters, etc.</param>
-        /// <param name="customHeaders">Optional headers.</param>
-        /// <param name="cancellationToken">Optional cancellation token.</param>
-        /// <returns></returns>
-        Task<HttpResponseMessage> SendRequestAsync(HttpMethod method, string relativeUrl, Dictionary<string, List<string>> customHeaders = null, CancellationToken cancellationToken = default(CancellationToken));
-
-        /// <summary>
-        /// Send a request with method <paramref name="method"/> to <paramref name="relativeUrl"/> with <paramref name="body"/>.
-        /// </summary>
-        /// <typeparam name="TBody">The type for the <paramref name="body"/>.</typeparam>
-        /// <param name="method">POST, GET, etc.</param>
-        /// <param name="relativeUrl">The Url relative to <see cref="BaseUri"/>, including parameters, etc.</param>
-        /// <param name="body">The body (content) for the request.</param>
-        /// <param name="customHeaders">Optional headers.</param>
-        /// <param name="cancellationToken">Optional cancellation token.</param>
-        /// <returns></returns>
-        Task<HttpResponseMessage> SendRequestAsync<TBody>(HttpMethod method, string relativeUrl, TBody body = default(TBody),
-            Dictionary<string, List<string>> customHeaders = null,
-            CancellationToken cancellationToken = default(CancellationToken));
-
-        /// <summary>
-        /// Send a request with method <paramref name="method"/> to <paramref name="relativeUrl"/> with <paramref name="body"/> and expect a result of a specific type.
-        /// </summary>
-        /// <typeparam name="TResponse">The type for the result.</typeparam>
-        /// <typeparam name="TBody">The type for the <paramref name="body"/>.</typeparam>
-        /// <param name="method">POST, GET, etc.</param>
-        /// <param name="relativeUrl">The Url relative to <see cref="BaseUri"/>, including parameters, etc.</param>
-        /// <param name="body">The body (content) for the request.</param>
-        /// <param name="customHeaders">Optional headers.</param>
-        /// <param name="cancellationToken">Optional cancellation token.</param>
-        /// <returns></returns>
-        Task<HttpOperationResponse<TResponse>> SendRequestAsync<TResponse, TBody>(HttpMethod method, string relativeUrl,
-            TBody body = default(TBody), Dictionary<string, List<string>> customHeaders = null,
-            CancellationToken cancellationToken = default(CancellationToken));
-
         #endregion
     }
 }
