+++ test/Nexus.Link.Libraries.Crud.Web.Test/RestClientTests/CrudManyToOneRecursiveRestClientTest.cs
@@ -28,7 +28,7 @@ namespace Nexus.Link.Libraries.Crud.Web.Test.RestClientTests
         {
             FulcrumApplicationHelper.UnitTestSetup(typeof(CrudManyToOneRecursiveRestClientTest).FullName);
             HttpClientMock = new Mock<IHttpClient>();
-            Libraries.Web.RestClientHelper.RestClient.HttpClient = HttpClientMock.Object;
+            Libraries.Web.RestClientHelper.HttpSender.HttpClient = HttpClientMock.Object;
             _parentChildrenClient = new CrudManyToOneRestClient<Person, Guid>(ResourcePath);
             _oneManyClient = new CrudManyToOneRestClient<Person, Guid>(ResourcePath, "One", "Many");
             _person = new Person()
