+++ test/Nexus.Link.Libraries.Crud.Web.Test/RestClientTests/CrudRestClientTests.cs
@@ -28,7 +28,7 @@ namespace Nexus.Link.Libraries.Crud.Web.Test.RestClientTests
         {
             FulcrumApplicationHelper.UnitTestSetup(typeof(CrudRestClientTests).FullName);
             HttpClientMock = new Mock<IHttpClient>();
-            Libraries.Web.RestClientHelper.RestClient.HttpClient = HttpClientMock.Object;
+            Libraries.Web.RestClientHelper.HttpSender.HttpClient = HttpClientMock.Object;
             _client = new CrudRestClient<Person, Guid>(ResourcePath);
             _person = new Person()
             {
